backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
email-validator>=2.0.0
groq>=0.36.0
python-dotenv==1.0.0
httpx>=0.24.0
PyPDF2>=3.0.0
google-generativeai>=0.3.0
pillow>=10.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
# Optional: OpenAI for video analysis fallback (paid API)
# openai>=1.0.0

